requirements.txt
pyelftools>=0.30

